Fixed the drawing mechanism.
The drawing was really wonky. The only reason it appeared to be working at times was because it was continuously being drawn due to a timer... no Idea why I thought that was the right way to do it. Now we are drawing it once and building an image out of it. It is being re-drawn every time the page is resized.